refactor: Add pages/auth folder
