+++ src/router/router.js
@@ -1,15 +1,14 @@
 import { createBrowserRouter } from 'react-router-dom';
 import DefaultLayout from '../layout/DefaultLayout';
 import Home from '../pages/Home';
-import Login from '../pages/Login/Login';
+import Login from '../pages/Auth/Login/Login';
 import Menu from '../pages/Menu';
-import Main from '../pages/Main';
 import Notice from '../pages/Notice/Notice';
 import CareDiary from '../pages/CareDiary';
 import Donation from '../pages/Donation/Donation';
 import DonationItemDetail from '../pages/Donation/DonationItemDetail';
 import CommonSense from '../pages/Notice/CommonSense';
-import Register from '../pages/Register/Register';
+import Register from '../pages/Auth/Register/Register';
 import DonationRecord from '../pages/DonationRecord';
 
 const routes = [
